requirements.txt
Django==5.0.3
djangorestframework
psycopg2
drf-spectacular
flake8
httpx
adrf